Possibly use Requires instead of If block
The check in the if block for security principle works, but there is a simpler way if you are already requiring PS 3.0 or higher. [Requires](https://technet.microsoft.com/en-us/library/hh847765.aspx) make this kind of check extremely easy.

No offense taken if you don't want to go this route.